Moves carouselv3 to carousel
