+++ src/Components/Artist/ArtistCollectionsRail/__tests__/ArtistCollectionsRail.test.tsx
@@ -1,6 +1,6 @@
 import { CollectionsRailFixture } from "Apps/__tests__/Fixtures/Collections"
 import { mockTracking } from "Artsy/Analytics"
-import { ArrowButton } from "Components/v2/CarouselV3"
+import { ArrowButton } from "Components/v2/Carousel"
 import { mount } from "enzyme"
 import "jest-styled-components"
 import { drop } from "lodash"
