+++ src/Components/v2/Carousel.tsx
@@ -1,234 +1,433 @@
-import { Box, ChevronIcon, Flex, media } from "@artsy/palette"
-import React, { ReactNode, useState } from "react"
-import Slick, { Settings } from "react-slick"
+import { ChevronIcon, color, Flex, space } from "@artsy/palette"
+import { Options as FlickityOptions } from "flickity"
+import React, { Fragment } from "react"
 import styled from "styled-components"
 import { left, LeftProps, right, RightProps } from "styled-system"
 import { Media } from "Utils/Responsive"
 
-const SLIDES_TO_SCROLL = 5
+/**
+ * Note: we have the commercial license, which allows us to use this in our MIT licensed codebase,
+ * but non-Artsy devs would technically be using the gplv3 version.
+ */
 
-type Arrow = (
-  props: {
-    currentSlideIndex: number
-    slidesToScroll?: number
-    Arrow: React.ReactType
-    getSlick: () => any
-  } & Props
-) => React.ReactNode
+interface CarouselProps {
+  /**
+   * This is designed to handle any shape of data passed, as long as its an array
+   */
+  data: any
 
-interface Props {
-  settings?: Settings
-  height?: number
-  data: object[] // This is designed to handle any shape of data passed, as long as its an array
-  render: (slide) => ReactNode
-  renderLeftArrow?: Arrow
-  renderRightArrow?: Arrow
+  /**
+   * Passes CarouselRef
+   */
+  setCarouselRef?: (CarouselRef) => void
+
+  /**
+   * If this carousel contains only one visible image on render set to true (for SSR)
+   */
+  oneSlideVisible?: boolean
+
+  /**
+   * The height of the carousel
+   */
+  height?: string
+
+  /**
+   * The width of the carousel
+   */
+  width?: string
+
+  /**
+   * Callback when forward / backward arrows are clicked
+   */
   onArrowClick?: () => void
+
+  /**
+   * The render callback returns an image to display
+   */
+  render: (slide) => React.ReactNode
+
+  /**
+   * Flickity options
+   */
+  options?: FlickityOptions
+
+  /**
+   * Show or hide arrows. Defaults to true
+   */
+  showArrows?: boolean
+
+  /**
+   * Custom renderer for the left arrow
+   */
+  renderLeftArrow?: ArrowProps
+
+  /**
+   * Custom renderer for the right arrow
+   */
+  renderRightArrow?: ArrowProps
 }
 
-export class Carousel extends React.Component<Props> {
+type ArrowProps = (props: {
+  currentSlideIndex: number
+  Arrow: React.ReactType
+  flickity: {
+    previous: (isWrapped?: boolean, isInstant?: boolean) => void
+    next: (isWrapped?: boolean, isInstant?: boolean) => void
+  }
+}) => React.ReactNode
+
+export class Carousel extends React.Component<CarouselProps> {
   static defaultProps = {
-    height: 300,
+    height: "300px",
+    oneSlideVisible: false,
   }
 
   render() {
     return (
       <>
-        <Media at="xs">
-          <SmallCarousel {...this.props} />
-        </Media>
         <Media greaterThan="xs">
           <LargeCarousel {...this.props} />
+        </Media>
+        <Media at="xs">
+          <SmallCarousel {...this.props} />
         </Media>
       </>
     )
   }
 }
 
-export const LargeCarousel = (props: Props) => {
-  const [currentSlideIndex, setSlideIndex] = useState(0)
-
-  const slickConfig = {
-    arrows: false,
-    draggable: false,
-    infinite: false,
-    speed: 500,
-    variableWidth: true,
-    slidesToScroll: SLIDES_TO_SCROLL,
-    afterChange: currentIndex => setSlideIndex(currentIndex),
-    ...props.settings,
-  }
-
-  let slickRef = null
-
-  const LeftArrow = () => {
-    return (
-      <ArrowButton
-        left={-8}
-        onClick={() => {
-          slickRef.slickPrev && slickRef.slickPrev() // check existence for tests
-          props.onArrowClick && props.onArrowClick()
-        }}
-      >
-        <ChevronIcon direction="left" fill="black100" width={30} height={30} />
-      </ArrowButton>
-    )
-  }
-
-  const RightArrow = () => {
-    return (
-      <ArrowButton
-        right={-8}
-        onClick={() => {
-          slickRef.slickNext && slickRef.slickNext() // check existence for tests
-          props.onArrowClick && props.onArrowClick()
-        }}
-      >
-        <ChevronIcon direction="right" fill="black100" width={30} height={30} />
-      </ArrowButton>
-    )
-  }
+export const LargeCarousel: React.FC<CarouselProps> = props => {
   return (
-    <Flex
-      flexDirection="row"
-      justifyContent="space-around"
-      alignItems="center"
-      height={props.height}
-    >
-      {props.renderLeftArrow ? (
-        props.renderLeftArrow({
-          Arrow: LeftArrow,
-          currentSlideIndex,
-          getSlick: () => slickRef,
-          ...props,
-        })
-      ) : (
-        <LeftArrow />
-      )}
-
-      <CarouselContainer height={props.height}>
-        <Slick {...slickConfig} ref={slider => (slickRef = slider)}>
-          {props.data.map((slide, index) => {
-            return <Box key={index}>{props.render(slide)}</Box>
-          })}
-        </Slick>
-      </CarouselContainer>
-
-      {props.renderRightArrow ? (
-        props.renderRightArrow({
-          Arrow: RightArrow,
-          currentSlideIndex,
-          slidesToScroll: SLIDES_TO_SCROLL,
-          getSlick: () => slickRef,
-          ...props,
-        })
-      ) : (
-        <RightArrow />
-      )}
-    </Flex>
+    <BaseCarousel
+      showArrows
+      options={{
+        cellAlign: "left",
+        contain: true,
+        draggable: false,
+        freeScroll: false,
+        groupCells: true,
+        lazyLoad: false,
+        pageDots: false,
+        wrapAround: false,
+        ...props.options,
+      }}
+      {...props}
+    />
   )
 }
 
-export const SmallCarousel = (props: Props) => {
-  const slickConfig = {
-    arrows: false,
-    dots: false,
-    speed: 300,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    swipeToSlide: true,
-    touchThreshold: 60,
-    variableWidth: true,
-    ...props.settings,
-  }
-
+export const SmallCarousel: React.FC<CarouselProps> = props => {
+  // Only render pageDots and enable draggable if more than one slide
+  const hasMultipleSlides = props.data.length > 1
   return (
-    <Flex justifyContent="space-around" alignItems="center">
-      <CarouselContainer height={props.height}>
-        <Slick {...slickConfig}>
-          {props.data.map((slide, index) => {
-            return <Box key={index}>{props.render(slide)}</Box>
-          })}
-        </Slick>
-      </CarouselContainer>
-    </Flex>
+    <BaseCarousel
+      showArrows={false}
+      options={{
+        cellAlign: "left",
+        draggable: hasMultipleSlides,
+        freeScroll: false,
+        contain: true,
+        friction: 0.3,
+        pageDots: hasMultipleSlides,
+        prevNextButtons: false,
+        wrapAround: false,
+        ...props.options,
+      }}
+      {...props}
+    />
   )
 }
 
-const CarouselContainer = styled.div<{ height?: number }>`
+interface BaseCarouselState {
+  currentSlideIndex: number
+  lastItemVisible: boolean
+  isMounted: boolean
+}
+
+export class BaseCarousel extends React.Component<
+  CarouselProps,
+  BaseCarouselState
+> {
+  state = {
+    currentSlideIndex: 0,
+    lastItemVisible: false,
+    isMounted: false,
+  }
+
+  /**
+   * A reference to the Flickity instance
+   */
+  flickity = null
+  carouselRef = null
+
+  /**
+   * Options to pass to underlying flickity component
+   */
+  options: FlickityOptions = {
+    lazyLoad: true,
+    prevNextButtons: false,
+  }
+
+  constructor(props) {
+    super(props)
+
+    // Flickity carousel options can be overriden via `props.options`
+    this.options = {
+      ...this.options,
+      ...props.options,
+    }
+  }
+
+  /**
+   * Since Flickity doesn't support SSR rendering, we need to load it once the
+   * client has mounted. During the server-side pass we use a Flex wrapper instead.
+   */
+  componentDidMount() {
+    const Flickity = require("flickity")
+    const { setCarouselRef } = this.props
+
+    this.flickity = new Flickity(this.carouselRef, this.options)
+
+    this.setState(
+      {
+        isMounted: true,
+      },
+      () => {
+        if (setCarouselRef) {
+          setCarouselRef(this.flickity)
+        }
+        this.flickity.on("select", this.handleSlideChange)
+      }
+    )
+  }
+
+  componentWillUnmount() {
+    if (this.flickity) {
+      this.flickity.off("select")
+    }
+  }
+
+  handleSlideChange = index => {
+    this.setState({
+      currentSlideIndex: index,
+      lastItemVisible: this.checkLastItemVisible(),
+    })
+  }
+
+  checkLastItemVisible = () => {
+    const lastItemVisible = this.flickity.selectedElements.includes(
+      this.flickity.getLastCell().element
+    )
+    return lastItemVisible
+  }
+
+  renderLeftArrow = () => {
+    const { onArrowClick, renderLeftArrow, showArrows } = this.props
+
+    if (!showArrows) {
+      return null
+    }
+
+    const showLeftArrow =
+      this.state.currentSlideIndex !== 0 || this.options.wrapAround === true
+
+    const Arrow = () => (
+      <ArrowWrapper left={-space(4)} showArrow={showLeftArrow}>
+        <ArrowButton
+          onClick={() => {
+            this.flickity.previous()
+
+            if (onArrowClick) {
+              onArrowClick()
+            }
+          }}
+        >
+          <ChevronIcon
+            direction="left"
+            fill="black100"
+            width={30}
+            height={30}
+          />
+        </ArrowButton>
+      </ArrowWrapper>
+    )
+
+    // Override default arrow display if custom render callback has been passed
+    if (renderLeftArrow) {
+      return renderLeftArrow({
+        Arrow,
+        currentSlideIndex: this.state.currentSlideIndex,
+        flickity: this.flickity,
+      })
+    }
+
+    return <Arrow />
+  }
+
+  renderRightArrow = () => {
+    const { onArrowClick, renderRightArrow, showArrows } = this.props
+
+    if (!showArrows) {
+      return null
+    }
+
+    const showRightArrow =
+      !this.state.lastItemVisible || this.options.wrapAround === true
+
+    const Arrow = () => (
+      <ArrowWrapper right={-space(4)} showArrow={showRightArrow}>
+        <ArrowButton
+          onClick={() => {
+            this.flickity.next()
+
+            if (onArrowClick) {
+              onArrowClick()
+            }
+          }}
+        >
+          <ChevronIcon
+            direction="right"
+            fill="black100"
+            width={30}
+            height={30}
+          />
+        </ArrowButton>
+      </ArrowWrapper>
+    )
+
+    // Override default arrow display if custom render callback has been passed
+    if (renderRightArrow) {
+      return renderRightArrow({
+        Arrow,
+        currentSlideIndex: this.state.currentSlideIndex,
+        flickity: this.flickity,
+      })
+    }
+
+    return <Arrow />
+  }
+
+  render() {
+    const { isMounted } = this.state
+    const { data, height, oneSlideVisible, render } = this.props
+
+    // FIXME: During SSR pass want to hide other images. Work around for lack
+    // of SSR support in Flickity.
+    const carouselImages =
+      typeof window === "undefined" && oneSlideVisible ? [data[0]] : data
+
+    return (
+      <>
+        <Flex
+          flexDirection="row"
+          position="relative"
+          justifyContent="space-around"
+          alignItems="center"
+          height={height}
+        >
+          {this.renderLeftArrow()}
+
+          <CarouselContainer height={height} isMounted={isMounted}>
+            <FlickityCarousel
+              isMounted={isMounted}
+              ref={c => (this.carouselRef = c)}
+            >
+              {carouselImages.map((slide, index) => {
+                return <Fragment key={index}>{render(slide)}</Fragment>
+              })}
+            </FlickityCarousel>
+          </CarouselContainer>
+
+          {this.renderRightArrow()}
+        </Flex>
+      </>
+    )
+  }
+}
+
+const FlickityCarousel = styled.div<{
+  isMounted: boolean
+}>`
+  display: ${props => (props.isMounted ? "block" : "flex")};
+`
+
+const CarouselContainer = styled.div<{
+  height?: string
+  isMounted: boolean
+}>`
   width: 100%;
+  position: relative;
+  overflow: ${props => (props.isMounted ? "visible" : "hidden")};
 
-  ${"" /*
-    FIXME: Revisit using react-slick -- too many hacks
-*/};
+  .flickity-viewport {
+    overflow: hidden;
+    width: 100%;
+  }
+
+  .flickity-slider {
+    img {
+      user-select: none;
+    }
+  }
+
+  .flickity-page-dots {
+    text-align: center;
+    height: 0;
+    padding-top: ${space(1)}px;
+
+    .dot {
+      width: 4px;
+      height: 4px;
+      border-radius: 100%;
+      display: inline-block;
+      margin: ${space(0.5)}px;
+      background-color: ${color("black10")};
+    }
+
+    .dot.is-selected {
+      background-color: ${color("black100")};
+    }
+  }
 
   ${props => {
     if (props.height) {
       return `
-        height: ${props.height}px;
-
-        .slick-slider, .slick-list, .slick-slide {
-          height: ${props.height}px;
-        }
+        height: ${props.height};
       `
     }
   }};
-
-  ${"" /*
-    FIXME: The below two rules are hacks for SSR to render properly.
-    Might have been fixed in https://github.com/artsy/reaction/pull/929
-  */};
-
-  .slick-track {
-    display: inline-flex;
-    width: 100% !important;
-  }
-
-  .slick-slide {
-    position: relative;
-    top: -18px;
-  }
-
-  ${"" /*
-    FIXME: On SSR mobile this shifts the image, must fix
-    Might be fixed in https://github.com/artsy/reaction/pull/929
-  */};
-
-  ${media.xs`
-    .slick-list {
-      padding: 0 !important;
-      height: ${p => p.height}px;
-    }
-  `};
-
-  .slick-dots li {
-    width: 0;
-
-    button {
-      &::before {
-        font-size: 4px;
-      }
-    }
-  }
 `
 
 export const ArrowButton = styled(Flex)<
-  LeftProps & RightProps & { height?: number }
+  LeftProps & RightProps & { height?: string }
 >`
   position: relative;
   cursor: pointer;
   display: flex;
   align-items: center;
   user-select: none;
-  opacity: 0.1;
+  opacity: 0.3;
 
   transition: opacity 0.25s;
-  min-height: ${p => p.height || 200}px;
+  min-height: ${p => p.height || "200px"};
 
   &:hover {
     opacity: 1;
   }
+`
 
+const ArrowWrapper = styled.div<{
+  left?: number
+  right?: number
+  showArrow: boolean
+}>`
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  min-width: 30px;
+  transition: opacity 0.25s;
+  opacity: ${props => (props.showArrow ? 1 : 0)};
+  pointer-events: ${props => (props.showArrow ? "auto" : "none")};
   ${left};
   ${right};
 `
