+++ src/Components/v2/__tests__/Carousel.test.tsx
@@ -1,7 +1,7 @@
 import { MockBoot } from "DevTools/MockBoot"
 import { mount } from "enzyme"
 import React from "react"
-import { Carousel, LargeCarousel, SmallCarousel } from "../CarouselV3"
+import { Carousel, LargeCarousel, SmallCarousel } from "../Carousel"
 
 describe("Carousel", () => {
   beforeAll(() => {
